# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mass Finder API",
        "description": "API para busca de paróquias e horários de missa",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{baseUrl}/api",
            "description": "API Base URL",
            "variables": {
                "baseUrl": {
                    "default": "http://localhost:8000"
                }
            }
        }
    ],
    "paths": {
...
        "/parish/mass-schedules": {
            "get": {
                "tags": [
                    "Parish"
                ],
                "summary": "Listar horários de missa da paróquia",
                "operationId": "dd4c6fa4dbc93975038fbdd9d5257dba",
                "responses": {
                    "200": {
                        "description": "Lista de horários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MassSchedule"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paróquia não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Parish"
                ],
                "summary": "Criar horário de missa",
                "operationId": "422335e1db0f714c5805033cfe11e131",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMassScheduleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Horário criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MassSchedule"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paróquia não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/parish/mass-schedules/{massSchedule}": {
            "put": {
                "tags": [
                    "Parish"
                ],
                "summary": "Atualizar horário de missa",
                "operationId": "6b64c2febaf6833837e3496b0b11bb73",
                "parameters": [
                    {
                        "name": "massSchedule",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMassScheduleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Horário atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MassSchedule"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Horário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Parish"
                ],
                "summary": "Remover horário de missa",
                "operationId": "ae300dd212d8668911df4483588cd25d",
                "parameters": [
                    {
                        "name": "massSchedule",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Horário removido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Horário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MassSchedule": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parish_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "day_of_week": {
                        "description": "0=Domingo, 1=Segunda, ..., 6=Sábado",
                        "type": "integer",
                        "maximum": 6,
                        "minimum": 0
                    },
                    "time": {
                        "type": "string",
                        "example": "08:00"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "CreateMassScheduleRequest": {
                "required": [
                    "day_of_week",
                    "time"
                ],
                "properties": {
                    "day_of_week": {
                        "description": "0=Domingo, 1=Segunda, ..., 6=Sábado",
                        "type": "integer",
                        "maximum": 6,
                        "minimum": 0
                    },
                    "time": {
                        "type": "string",
                        "pattern": "^([01]?[0-9]|2[0-3]):[0-5][0-9]$",
                        "example": "08:00"
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "GenericMessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operação realizada com sucesso."
                    }
                },
                "type": "object"
            },
            "ErrorValidationResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido via POST /api/auth/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}